+++ src/main/java/th/mfu/Repositories/StaffStudentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import th.mfu.Domain.StaffStudent;
 
-public interface StaffStudentRepository extends CrudRepository<StaffStudent, Integer> {
+public interface StaffStudentRepository extends CrudRepository<StaffStudent, Long> {
     
     StaffStudent findbyStaffId(String phonenumb);
 }
