+++ src/main/java/th/mfu/Repositories/VehicleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import th.mfu.Domain.Vehicle;
 
-public interface VehicleRepository extends CrudRepository<Vehicle, Integer> {
+public interface VehicleRepository extends CrudRepository<Vehicle, Long> {
     
     
 }
